workflows/basic/requirements.txt
lightgbm
matplotlib
scikit-learn
azureml-mlflow 
azureml-dataprep
